requirements.txt
streamlit==1.10.0
pandas==1.4.3
numpy==1.23.1
pillow==8.4.0
scikit_learn==0.24.2
altair==4.2.0
xgboost==1.3.3